+++ webots_ros2_turtlebot/launch/robot_launch.py
@@ -24,15 +24,21 @@ from launch.substitutions.path_join_substitution import PathJoinSubstitution
 from launch import LaunchDescription
 from launch_ros.actions import Node
 import launch
-from ament_index_python.packages import get_package_share_directory
+from ament_index_python.packages import get_package_share_directory, get_packages_with_prefixes
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.actions import IncludeLaunchDescription
 from webots_ros2_driver.webots_launcher import WebotsLauncher
 from webots_ros2_driver.utils import controller_url_prefix
 
 
 def get_ros2_nodes(*args):
     package_dir = get_package_share_directory('webots_ros2_turtlebot')
+    use_nav = LaunchConfiguration('nav', default=False)
+    use_slam = LaunchConfiguration('slam', default=False)
     robot_description = pathlib.Path(os.path.join(package_dir, 'resource', 'turtlebot_webots.urdf')).read_text()
     ros2_control_params = os.path.join(package_dir, 'resource', 'ros2control.yml')
+    nav2_params = os.path.join(package_dir, 'resource', 'nav2_params.yaml')
+    nav2_map = os.path.join(package_dir, 'resource', 'turtlebot3_burger_example_map.yaml')
     use_sim_time = LaunchConfiguration('use_sim_time', default=True)
 
     # TODO: Revert once the https://github.com/ros-controls/ros2_control/pull/444 PR gets into the release
@@ -91,21 +97,61 @@ def get_ros2_nodes(*args):
         arguments=['0', '0', '0', '0', '0', '0', 'base_link', 'base_footprint'],
     )
 
+    nav_nodes = []
+    # Navigation
+    os.environ['TURTLEBOT3_MODEL'] = 'burger'
+    if 'turtlebot3_navigation2' in get_packages_with_prefixes():
+        turtlebot_navigation = IncludeLaunchDescription(
+            PythonLaunchDescriptionSource(os.path.join(
+                get_package_share_directory('turtlebot3_navigation2'), 'launch', 'navigation2.launch.py')),
+            launch_arguments=[
+                ('map', nav2_map),
+                ('params_file', nav2_params),
+                ('use_sim_time', use_sim_time),
+            ],
+            condition=launch.conditions.IfCondition(use_nav))
+        nav_nodes.append(turtlebot_navigation)
+
+    # SLAM
+    if 'turtlebot3_cartographer' in get_packages_with_prefixes():
+        turtlebot_slam = IncludeLaunchDescription(
+            PythonLaunchDescriptionSource(os.path.join(
+                get_package_share_directory('turtlebot3_cartographer'), 'launch', 'cartographer.launch.py')),
+            launch_arguments=[
+                ('use_sim_time', use_sim_time),
+            ],
+            condition=launch.conditions.IfCondition(use_slam))
+        nav_nodes.append(turtlebot_slam)
+
+    # Wait for the simulation to be ready to start navigation nodes
+    nav_handler = []
+    if nav_nodes:
+        nav_handler.append(
+            launch.actions.RegisterEventHandler(
+                event_handler=launch.event_handlers.OnProcessExit(
+                    target_action=diffdrive_controller_spawner,
+                    on_exit=nav_nodes
+                )
+            )
+        )
+
     return [
         joint_state_broadcaster_spawner,
         diffdrive_controller_spawner,
         robot_state_publisher,
         turtlebot_driver,
         footprint_publisher,
-    ]
+    ] + nav_handler
 
 
 def generate_launch_description():
     package_dir = get_package_share_directory('webots_ros2_turtlebot')
     world = LaunchConfiguration('world')
+    mode = LaunchConfiguration('mode')
 
     webots = WebotsLauncher(
         world=PathJoinSubstitution([package_dir, 'worlds', world]),
+        mode=mode,
         ros2_supervisor=True
     )
 
@@ -123,6 +169,11 @@ def generate_launch_description():
             'world',
             default_value='turtlebot3_burger_example.wbt',
             description='Choose one of the world files from `/webots_ros2_turtlebot/world` directory'
+        ),
+        DeclareLaunchArgument(
+            'mode',
+            default_value='realtime',
+            description='Webots startup mode'
         ),
         webots,
         webots._supervisor,