+++ webots_ros2/CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog for package webots_ros2
 * Added Emitter and Receiver support in webots_ros2_driver
 * Changed undefined Lidar frequency to the default from the .proto file
 * Added Compass support in webots_ros2_driver
+* Added startup of the Turlebot navigation and mapping tools from the launch file.
 
 2023.0.2 (2023-02-07)
 ------------------